+++ src/main/java/com/caoccao/javet/interop/converters/JavetProxyConverter.java
@@ -21,7 +21,7 @@ import com.caoccao.javet.annotations.CheckReturnValue;
 import com.caoccao.javet.exceptions.JavetException;
 import com.caoccao.javet.interop.V8Runtime;
 import com.caoccao.javet.interop.callback.JavetCallbackContext;
-import com.caoccao.javet.interop.proxy.JavetUniversalInterceptionProxyHandler;
+import com.caoccao.javet.interop.proxy.JavetUniversalProxyHandler;
 import com.caoccao.javet.values.V8Value;
 import com.caoccao.javet.values.reference.IV8ValueObject;
 import com.caoccao.javet.values.reference.V8ValueProxy;
@@ -37,42 +37,12 @@ import java.util.List;
 public class JavetProxyConverter extends JavetObjectConverter {
 
     /**
-     * The constant PROXY_TARGET.
-     *
-     * @since 0.9.6
-     */
-    protected static final String PROXY_TARGET = "target";
-
-    /**
      * Instantiates a new Javet proxy converter.
      *
      * @since 0.9.6
      */
     public JavetProxyConverter() {
         super();
-    }
-
-    @Override
-    protected Object toObject(V8Value v8Value, final int depth) throws JavetException {
-        validateDepth(depth);
-        if (v8Value instanceof V8ValueProxy) {
-            V8ValueProxy v8ValueProxy = (V8ValueProxy) v8Value;
-            try (IV8ValueObject iV8ValueObjectHandler = v8ValueProxy.getHandler()) {
-                Long handle = iV8ValueObjectHandler.getLong(PROXY_TARGET);
-                if (handle != null) {
-                    JavetCallbackContext javetCallbackContext = v8ValueProxy.getV8Runtime().getCallbackContext(handle);
-                    if (javetCallbackContext != null) {
-                        JavetUniversalInterceptionProxyHandler<Object> javetUniversalInterceptionProxyHandler =
-                                (JavetUniversalInterceptionProxyHandler<Object>) javetCallbackContext.getCallbackReceiver();
-                        Object returnObject = javetUniversalInterceptionProxyHandler.getTargetObject();
-                        if (returnObject != null) {
-                            return returnObject;
-                        }
-                    }
-                }
-            }
-        }
-        return super.toObject(v8Value, depth);
     }
 
     @Override
@@ -85,10 +55,9 @@ public class JavetProxyConverter extends JavetObjectConverter {
         }
         V8ValueProxy v8ValueProxy = v8Runtime.createV8ValueProxy();
         try (IV8ValueObject iV8ValueObjectHandler = v8ValueProxy.getHandler()) {
-            JavetUniversalInterceptionProxyHandler<Object> javetUniversalInterceptionProxyHandler =
-                    new JavetUniversalInterceptionProxyHandler<>(object);
-            List<JavetCallbackContext> javetCallbackContexts = iV8ValueObjectHandler.bind(
-                    javetUniversalInterceptionProxyHandler);
+            JavetUniversalProxyHandler<Object> javetUniversalProxyHandler =
+                    new JavetUniversalProxyHandler<>(v8Runtime, object);
+            List<JavetCallbackContext> javetCallbackContexts = iV8ValueObjectHandler.bind(javetUniversalProxyHandler);
             iV8ValueObjectHandler.set(PROXY_TARGET, javetCallbackContexts.get(0).getHandle());
         }
         v8Value = v8ValueProxy;