+++ docs/release_notes.rst
@@ -8,7 +8,7 @@ Release Notes
 * Added ``IV8ValueProxy`` and ``V8ValueProxy``
 * Added ``getTarget()``, ``getHandler()``, ``isRevoked()`` and ``revoke()`` to ``IV8ValueProxy``
 * Added ``createV8ValueProxy()`` to ``V8Runtime``
-* Added ``JavetUniversalInterceptionProxyHandler`` and ``JavetProxyConverter``
+* Added ``JavetUniversalProxyHandler`` and ``JavetProxyConverter``
 
 0.9.5
 -----